extract some config to be shared
