+++ hw12_13_14_15_calendar/cmd/calendar/main.go
@@ -19,6 +19,7 @@ import (
 	pbEventV1 "github.com/dima-study/otus2405/hw12_13_14_15_calendar/internal/api/proto/event/v1"
 	calendarBusiness "github.com/dima-study/otus2405/hw12_13_14_15_calendar/internal/business/calendar"
 	helloBusiness "github.com/dima-study/otus2405/hw12_13_14_15_calendar/internal/business/hello"
+	"github.com/dima-study/otus2405/hw12_13_14_15_calendar/internal/config"
 	"github.com/dima-study/otus2405/hw12_13_14_15_calendar/internal/grpc/gw"
 	internalhttp "github.com/dima-study/otus2405/hw12_13_14_15_calendar/internal/http"
 	httpMiddleware "github.com/dima-study/otus2405/hw12_13_14_15_calendar/internal/http/middleware"
@@ -188,10 +189,10 @@ func run(ctx context.Context, logger *slog.Logger, levelVar *slog.LevelVar) erro
 
 func initStorage(cfg Config) (model.Storage, func() error, error) {
 	switch cfg.EventStorageType {
-	case EventStorageTypeMemory:
+	case config.EventStorageTypeMemory:
 		storage := memoryStorage.NewStorage()
 		return storage, func() error { return nil }, nil
-	case EventStorageTypePg:
+	case config.EventStorageTypePg:
 		storage, err := pgStorage.NewStorage(cfg.EventStoragePg.DataSource)
 		if err != nil {
 			return nil, nil, err
